nothing

+++ delshort.py
@@ -1,15 +1,18 @@
 #!/data/data/com.termux/files/usr/bin/env python
 from pathlib import Path
 from sys import exit
+
 from dh import is_binary
 from fastwalk import walk_files
 
 
 def process_file(fp) -> None:
-    if fp.stat().st_size < 50 and len(fp.read_text().splitlines()) < 3:
+    if not fp.exists():
+        return
+    if fp.exists() and fp.stat().st_size < 50 and len(fp.read_text().splitlines()) < 3:
         fp.unlink()
         print(f"{fp.name} removed")
-    if len(fp.read_text().splitlines()) < 2:
+    if fp.exists() and len(fp.read_text().splitlines()) < 2 and fp.suffix not in {".js", ".min.js", ".css", ".min.css"}:
         fp.unlink()
         print(f"{fp.name} removed")
 
@@ -17,7 +20,7 @@ def process_file(fp) -> None:
 def main() -> None:
     for pth in walk_files("."):
         path = Path(pth)
-        if not is_binary(path):
+        if not is_binary(path) and path.exists():
             process_file(path)
         else:
             print(f"{path.name} is binary")
